feat: slippage extract to local storage
