+++ libs/shared/providers/src/redeemer/components/Redeemer.tsx
@@ -22,11 +22,11 @@ import { useIntl } from 'react-intl';
 import { useAccount, useBalance } from 'wagmi';
 
 import { usePrices } from '../../prices';
-import { PriceTolerancePopover, useSlippage } from '../../slippage';
 import { ConnectedButton } from '../../wagmi';
 import { useHandleAmountInChange, useHandleRedeem } from '../hooks';
 import { RedeemProvider, useRedeemState } from '../state';
 import { RedeemRoute } from './RedeemRoute';
+import { SettingsPopover } from './SettingsPopover';
 
 import type { BoxProps, ButtonProps, StackProps } from '@mui/material';
 import type { MouseEvent } from 'react';
@@ -60,7 +60,6 @@ function RedeemerWrapped({
   ...rest
 }: Omit<RedeemerProps, 'trackEvent' | 'tokenIn' | 'vaultContract'>) {
   const intl = useIntl();
-  const { value: slippage, set: setSlippage } = useSlippage();
   const { address, isConnected } = useAccount();
   const [anchorEl, setAnchorEl] = useState<HTMLButtonElement | null>(null);
   const [
@@ -86,14 +85,6 @@ function RedeemerWrapped({
   const handleSettingClick = (evt: MouseEvent<HTMLButtonElement>) => {
     setAnchorEl(evt.currentTarget);
     trackEvent({ name: 'open_settings' });
-  };
-
-  const handleSlippageChange = (val: number) => {
-    setSlippage(val);
-    trackEvent({
-      name: 'change_price_tolerance',
-      price_tolerance: val,
-    });
   };
 
   const redeemButtonLabel =
@@ -134,28 +125,10 @@ function RedeemerWrapped({
                 >
                   <FaGearComplexRegular />
                 </IconButton>
-                <PriceTolerancePopover
+                <SettingsPopover
                   open={!!anchorEl}
                   anchorEl={anchorEl}
                   onClose={() => setAnchorEl(null)}
-                  slippage={slippage}
-                  onSlippageChange={handleSlippageChange}
-                  buttonProps={{
-                    sx: {
-                      color: 'text.primary',
-                      background:
-                        'linear-gradient(90deg, #8C66FC 0%, #0274F1 100%)',
-                      '&:hover': {
-                        background:
-                          'linear-gradient(0deg, rgba(0, 0, 0, 0.20) 0%, rgba(0, 0, 0, 0.20) 100%), linear-gradient(90deg, #8C66FC 0%, #0274F1 100%)',
-                        opacity: 1,
-                      },
-                      '&:disabled': {
-                        opacity: 0.5,
-                        color: 'text.primary',
-                      },
-                    },
-                  }}
                 />
               </Stack>
             }
